refactor: group request classes by steam interface
move request classes into per-interface subnamespaces
(http\requests\isteamuser, isteamuserstats, iplayerservice)
to keep the requests directory manageable as more endpoints
are added. mirror the layout in the test suite and update
imports across tests and the readme.
